✨ feat: add profile dashboard layout
