+++ frontend/src/main.jsx
@@ -2,7 +2,7 @@ import EventCreationPage from '@/pages/EventCreationPage';
 import EventListingPage from '@/pages/EventListingPage';
 import HelpRequestPage from '@/pages/HelpRequestPage';
 import HomePage from '@/pages/HomePage.jsx';
-import ProfileManagement from '@/pages/ProfileManagementPage';
+import ProfileDashboard from '@/pages/ProfileDashboard';
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
 import { StrictMode } from 'react';
 import { createRoot } from 'react-dom/client';
@@ -14,6 +14,8 @@ import MainLayout from './layouts/MainLayout';
 import HelpRequestDetailsPage from './pages/HelpRequestDetailsPage';
 import MyEventsPage from './pages/MyEventsPage';
 import MyParticipationEvents from './pages/MyParticipationEvents';
+import EditProfile from './pages/profile/EditProfile';
+import VolunteerHistory from './pages/profile/VolunteerHistory';
 import AuthProvider from './providers/AuthProvider';
 
 const queryClient = new QueryClient();
@@ -28,7 +30,13 @@ createRoot(document.getElementById('root')).render(
 							<Route index element={<HomePage />} />
 							<Route path="sign-in" element={<SignIn />} />
 							<Route path="sign-up" element={<SignUp />} />
-							<Route path="profile" element={<ProfileManagement />} />
+							<Route path="profile" element={<ProfileDashboard />}>
+								<Route index element={<EditProfile />} />
+								<Route path="edit-profile" element={<EditProfile />} />
+								<Route path="history" element={<VolunteerHistory />} />
+								<Route path="impact" element={<h1>impact</h1>} />
+								<Route path="*" element={<h1>404</h1>} />
+							</Route>
 							<Route path="my-events" element={<MyEventsPage />} />
 							<Route
 								path="my-participation-events"
