+++ frontend/src/pages/ProfileDashboard.jsx
@@ -1,0 +1,13 @@
+import Sidebar from '@/components/profile/Sidebar';
+import { Outlet } from 'react-router';
+
+export default function ProfileDashboard() {
+	return (
+		<div className="flex min-h-screen bg-gray-100">
+			<Sidebar />
+			<main className="flex-1 p-8">
+				<Outlet />
+			</main>
+		</div>
+	);
+}
